# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar MarketPay API",
    "version": "1.0.0",
    "description": "Backend API for Stellar MarketPay - A decentralized freelance marketplace built on the Stellar blockchain.\n\n## Authentication\nMost endpoints require JWT authentication obtained via the **/api/auth** challenge-response flow using Stellar wallet signatures.\n\n## Pagination\nList endpoints support cursor-based pagination via the `after` query parameter. Responses include `next_cursor` and `has_more` fields.",
    "contact": {
      "name": "Stellar MarketPay Team",
      "email": "support@stellarmarketpay.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.stellarmarketpay.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth (Stellar wallet signature)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT cookie set after authentication"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Developer API key for public API endpoints"
      }
    },
...
  },
  "paths": {
    "/api/webauthn/register-options": {
      "post": {
        "summary": "Get WebAuthn registration options",
        "tags": [
          "WebAuthn"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Registration options for creating a passkey"
          }
        }
      }
    },
    "/api/webauthn/register-verify": {
      "post": {
        "summary": "Verify and store a WebAuthn credential",
        "tags": [
          "WebAuthn"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "credential"
                ],
                "properties": {
                  "credential": {
                    "type": "object"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Passkey registered"
          },
          "400": {
            "description": "Verification failed or no pending challenge"
          }
        }
      }
    },
    "/api/webauthn/login-options": {
      "post": {
        "summary": "Get WebAuthn login options",
        "tags": [
          "WebAuthn"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "publicKey"
                ],
                "properties": {
                  "publicKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication options"
          }
        }
      }
    },
    "/api/webauthn/login-verify": {
      "post": {
        "summary": "Verify WebAuthn authentication and issue JWT",
        "tags": [
          "WebAuthn"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "credential",
                  "publicKey"
                ],
                "properties": {
                  "credential": {
                    "type": "object"
                  },
                  "publicKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token issued"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/api/webauthn/credentials": {
      "get": {
        "summary": "List registered passkeys",
        "tags": [
          "WebAuthn"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Credential list"
          }
        }
      }
    },
    "/api/webauthn/credentials/{id}": {
      "delete": {
        "summary": "Remove a passkey",
        "tags": [
          "WebAuthn"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Passkey removed"
          }
        }
      }
    },
    "/api/webauthn/admin/credentials": {
      "get": {
        "summary": "Admin list credentials for any user",
        "tags": [
          "WebAuthn"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "publicKey",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential list"
          }
        }
      }
    },
    "/api/webauthn/admin/credentials/{id}": {
      "delete": {
        "summary": "Admin revoke a passkey",
        "tags": [
          "WebAuthn"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Passkey revoked"
          }
        }
      }
    },
...
  }
}